Add fixed height to header container
Fixes WEB-1778. The problem is that as you scroll down and get to the
point where the header switches from full to sticky, the page jumps up
as well. This is because some of header elements become absolutely
positioned, taking them out of the height calculation for the header
container and causing the now-offscreen container to shrink. This in turn
causes the page to jump as the content shifts up to fill the
vacuum. Assigning the header container a fixed height keeps the page
from jumping and any visual side-effects are offstage.

This is only an issue at 769px and above because below that we have a
static mobile header.